renamed protocard to standard card
