+++ app/frontend/src/components/Cards/StandardCard.tsx
