+++ app/frontend/src/components/Cards/_StandardCard.scss
