+++ app/frontend/src/components/Cards/_index.scss
@@ -1,2 +1,2 @@
-@forward "ProtoCard";
+@forward "StandardCard";
 @forward "CircleCard";
